Fix all warnings on MS VS, also fix endian bug.
Most warnings are implicit conversions, inspecting the warnings led to finding a bug in the endian-shim where htobe32 was incorrectly defined as htons instead of htonl

+++ contrib/pthread-shim.c
@@ -124,7 +124,7 @@ int pthread_cond_timedwait_broken(pthread_cond_t *cond, pthread_mutex_t *mutex,
 		if (RIST_UNLIKELY(duration < 0)) {
 			duration = 1;
 		}
-		ms = duration;
+		ms = (DWORD)duration;
 	}
 
 	if (!SleepConditionVariableCS(cond, mutex, ms)) {
